requirements.txt
django==4.0.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
requests==2.28.0
djangorestframework-simplejwt==4.6.0